# expense-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Expense Management API",
        "description": "Multi-Tenant SaaS Expense Management API — secure, role-based, company-scoped.",
        "contact": {
            "email": "gplanet.tech@gmail.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Local / Development"
        },
        {
            "url": "https://api.yourdomain.com/api",
            "description": "Production"
        }
    ],
    "paths": {
        "/audit-logs": {
            "get": {
                "tags": [
                    "Audit Logs"
                ],
                "summary": "List audit logs (Admin only, company-scoped, filterable)",
                "operationId": "26a564983f3075fc795357c172b819f1",
                "parameters": [
                    {
                        "name": "action",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "create",
                                "update",
                                "delete"
                            ]
                        }
                    },
                    {
                        "name": "model_type",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "Expense"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 25
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated audit log list"
                    },
                    "403": {
                        "description": "Admin role required"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/audit-logs/{id}": {
            "get": {
                "tags": [
                    "Audit Logs"
                ],
                "summary": "Get a single audit log entry (Admin only, company-scoped)",
                "operationId": "dfac862e27ff32ccc892f3807e86ac90",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Audit log detail"
                    },
                    "403": {
                        "description": "Admin role required"
                    },
                    "404": {
                        "description": "Not found or cross-company"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new company and its first Admin user",
                "operationId": "ed42f392010c301892e640067141676e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "company_name",
                                    "company_email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jane Smith"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jane@acme.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secret123",
                                        "minLength": 8
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "secret123"
                                    },
                                    "company_name": {
                                        "type": "string",
                                        "example": "Acme Corp"
                                    },
                                    "company_email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "info@acme.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registration successful — token returned"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login and receive a Bearer token",
                "operationId": "8cb7742ae98d0990acb2907c4e00bf24",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful — token returned"
                    },
                    "422": {
                        "description": "Wrong credentials"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Revoke the current access token",
                "operationId": "0d29c740dc06d77602af8095bf4654cc",
                "responses": {
                    "200": {
                        "description": "Logged out"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/auth/user": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get the authenticated user's profile",
                "operationId": "a77c8027188561af13809580ef5af170",
                "responses": {
                    "200": {
                        "description": "Profile retrieved"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Sanctum token — obtain via /auth/login or /auth/register",
                "bearerFormat": "token",
                "scheme": "bearer"
            }
        }
    },
...
}